Technician list

+++ src/components/clients/clients.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import withRouter from 'react-router-dom/es/withRouter';
+import { withRouter, Redirect } from 'react-router-dom';
 
 import ClientList from './client-list';
 
@@ -10,6 +10,7 @@ import * as actions from '../../actions';
 
 const Clients = props => {
   if (!props.clientsList || !props.me) return <div className="loader" />;
+  if (props.me.type === 'client') return <Redirect to={`${routes.webRoot}/dashboard`} />;
 
   const onEdit = client => props.history.push(`${routes.webRoot}/clients/${client.id}/edit`);
   const onDelete = client => props.deleteUser(client);
